refactor(SDK-1118): nest onboarding Contractor PaymentMethod under onboarding/
Aligns Contractor/PaymentMethod with the Employee domain's onboarding/management/shared
convention, where the older onboarding component previously sat at the domain root
instead of alongside its new management/shared siblings. Pure move; no behavior change.
