+++ src/components/Contractor/OnboardingFlow/OnboardingFlowComponents.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from 'react-i18next'
 import { ContractorList } from '../ContractorList/onboarding/ContractorList'
 import { ContractorProfile } from '../Profile'
 import { Address } from '../Address'
-import { PaymentMethod } from '../PaymentMethod/PaymentMethod'
+import { PaymentMethod } from '../PaymentMethod/onboarding/PaymentMethod'
 import { NewHireReport } from '../NewHireReport/NewHireReport'
 import { ContractorSubmit } from '../Submit/Submit'
 import type { ContractorDetailsFormData } from '../Profile/shared'
