+++ src/components/Contractor/PaymentMethod/onboarding/PaymentMethod.test.tsx
