+++ src/components/Contractor/PaymentMethod/onboarding/types.ts
