+++ src/components/Contractor/SelfOnboardingFlow/SelfOnboardingComponents.tsx
@@ -5,7 +5,7 @@ import type { BaseComponentInterface } from '@/components/Base'
 import { Landing as LandingComponent } from '@/components/Contractor/Landing'
 import { ContractorProfile as ProfileComponent } from '@/components/Contractor/Profile'
 import { Address as AddressComponent } from '@/components/Contractor/Address'
-import { PaymentMethod as PaymentMethodComponent } from '@/components/Contractor/PaymentMethod/PaymentMethod'
+import { PaymentMethod as PaymentMethodComponent } from '@/components/Contractor/PaymentMethod/onboarding/PaymentMethod'
 import { DocumentSigner as DocumentSignerComponent } from '@/components/Contractor/Documents/DocumentSigner'
 import { OnboardingSummary as OnboardingSummaryComponent } from '@/components/Contractor/OnboardingSummary'
