+++ src/components/Contractor/exports/contractorOnboarding.ts
@@ -22,8 +22,8 @@ export {
 export type { ContractorDetailsFormData as ContractorProfileFormData } from '../Profile/shared'
 export { Address, type AddressProps } from '../Address'
 export type { AddressDefaultValues } from '../Address/types'
-export { PaymentMethod } from '../PaymentMethod/PaymentMethod'
-export type { PaymentMethodProps } from '../PaymentMethod/types'
+export { PaymentMethod } from '../PaymentMethod/onboarding/PaymentMethod'
+export type { PaymentMethodProps } from '../PaymentMethod/onboarding/types'
 export { NewHireReport } from '../NewHireReport/NewHireReport'
 export type { NewHireReportProps } from '../NewHireReport/types'
 export { ContractorSubmit, type ContractorSubmitProps } from '../Submit/Submit'
